Woeptidoo,
Everything works now, the meetings are successfully stored in the Neo4J
database, with the correct stuff and when the stuff is called for again,
it won't be duplicated. The next thing that I want to do is to improve the
speed a little bit by reducing the calls to the neo4j database.

After which I'm going to create a custom view for this, and store the
google auth in a session cookie, instead of working on the callback page.
:3